change commit

+++ Database/Procedures.sql
@@ -126,7 +126,7 @@ go
 create procedure deleteUsers
 as
 begin
-delete from UserTable 
+delete from UserTable where UserId > 0
 end;
 go
